+++ lib/puppet/provider/reboot/linux.rb
@@ -1,3 +1,6 @@
+require 'pathname'
+require Pathname.new(__FILE__).dirname + '../../../' + 'puppet/provider/reboot/posix'
+
 Puppet::Type.type(:reboot).provide :linux, :parent => :posix do
   confine :kernel => :linux